refactor: split daemon client facade
